Removing extended UIButton "CellButton"
Most of the work on this has been to remove my extended UIButton which
was called “CellButton.” I had the view listening via KVO to changes in
the model. This is changed now to just use UIButtons and the
viewController is listening for changes in the model and updating the
corresponding UIButton.

Also Timer updates.